requirements.txt
setuptools==65.6.3
wandb
matplotlib
numpy
joblib
termcolor
h5py
Pillow
opencv-python
imageio
imageio[ffmpeg]
imageio[pyav]

hydra-core
torch
torchvision
dm-env==1.6
dm_control==1.0.14
mujoco==2.3.7
pyopengl
pyopengl_accelerate
